% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easieRstations-package.R
\docType{package}
\name{easieRstations-package}
\alias{easieRstations}
\alias{easieRstations-package}
\title{easieRstations: Easy selection of weather stations using a shiny app and leaflet}
\description{
Provides a wEAther Station pIckER (EASIER) for 'NCEI' Global
    Summary of the Day, 'GSOD', and Australian Bureau of Meteorology, 'BOM',
    weather stations through a 'leaflet' driven 'shiny' interface to automate
    the download of selected weather stations' data using 'GSODR' or 'bomrang'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adamhsparks.github.io/easieRstations/}
  \item Report bugs at \url{https://github.com/adamhsparks/easieRstations/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

}
\keyword{internal}
